PJ: fix loading/saving Tsettings_file structure.
The old Tfastflush/Tslowflush union was done to save stack space.
However, the separation between the fields in Tfastflush could
potentially be different to those in Tsettings_file, leading to a
corrupt file.

Loading a corrupt file can result in setting buf->vs.zoomscale = 0,
which leads to a crash further on.